+++ OnePy/strategy.py
@@ -11,6 +11,8 @@ class StrategyBase(with_metaclass(MetaParams, object)):
     def __init__(self,marketevent):
 
         m = marketevent
+        self.pricetype = 'open'             # 控制计算的价格，可以再OnePy中用 set_pricetype控制
+
         self.bar = m.cur_bar_list
         self.bar_dict = m.bar_dict
         self.data = m.cur_bar_list[0]
@@ -21,10 +23,12 @@ class StrategyBase(with_metaclass(MetaParams, object)):
         self.margin = [i['margin'] for i in m.fill.margin_dict[m.instrument]]
         self.total = [i['total'] for i in m.fill.total_list]
         self.avg_price = [i['avg_price'] for i in m.fill.avg_price_dict[m.instrument]]
-
         self.unre_profit = [i['unre_profit'] for i in m.fill.unre_profit_dict[m.instrument]]
         self.re_profit = [i['re_profit'] for i in m.fill.re_profit_dict[m.instrument]]
-        self.pricetype = 'open' # 控制计算的价格，可以再OnePy中用 set_pricetype控制
+
+        self.indicator = marketevent.indicator
+        self.i = self.indicator     # shortcut
+
 
     def pips(self,n):
         n = n*1.0/self._mult